requirements.txt
flask
langchain
langchain-core
langchain-groq
python-dotenv
art
gitpython